Restructured and organized repository tree. Added README.md files.
