+++ Coding_Challenges/Org_Management/README.md
@@ -1,0 +1,3 @@
+# Org Management
+
+This was an idea that was put on a pause with no end date. The goal here was to track all devices in an organization and alert the organization/lab manager when a device was deleted from the organization.
