+++ Coding_Challenges/Org_Management/inv_gather_devices.py
