+++ Coding_Challenges/README.md
@@ -1,0 +1,3 @@
+# Coding Challenges
+
+The files here have been made as either a mental or coding exercise. Some of these are simply a proof of concept. Feel free to use as a template or module to import.
