+++ Coding_Challenges/print_24_hours.py
