+++ Coding_Challenges/remove_old.py
