+++ Coding_Challenges/trigger_lockdown.py
