+++ Customer_Projects/README.md
@@ -1,0 +1,3 @@
+# Customer_Projects
+
+These are scripts that have been used for deals or trials.
