+++ Demo_Org/README.md
@@ -1,0 +1,6 @@
+# Demo Org
+
+These are scripts that were used during the creation and testing of the finalized, private script that is managing the current Command Demo organization. These are not the actual scripts in use to manage the Demo environment.
+
+> [!WARNING]
+> Please use caution when planning to use these scripts as they may harm your organization if the documentation in the script is not fully read.
