+++ Demo_Org/create_trash_pois.py
