+++ QoL/custom_exceptions.py
