+++ VCE/README.md
@@ -1,0 +1,12 @@
+# VCE
+
+The scripts here are to help streamline the running of the VCE program by automating tedious and repetitive tasks.
+
+## End of the Event
+
+### Deleting Devices
+
+The [delete_devices](https://github.com/ian-young/API_Scripts/VCE/delete_devices.py) will delete all devices to aid with the cleanup of the organization after the event.
+
+> [!WARNING]
+> Please use caution if planning to use these scripts as they will harm your organization.
