+++ VCE/gather_devices.py
